further compute and consume cohorts
