+++ src/components/BarChart.vue
@@ -4,20 +4,8 @@
 
 <script>
 import Chart from "chart.js";
+import { getNestedValue } from "@/helpers";
 
-function getNestedValue(dottedKey, obj) {
-  let keys = dottedKey.split(".");
-  let out = obj;
-  let key;
-  for (let i = 0; i < keys.length; i++) {
-    key = keys[i];
-    if (out[key] === undefined) {
-      return undefined;
-    }
-    out = out[key];
-  }
-  return out;
-}
 export default {
   name: "BarChart",
   props: {
